Remove dual collection use from import_calibs.py.
The original DC2 script had a user-specified chained collection and a
hardcoded calibration collection that had to be a member of the chained
collection. This change replaces the system with a single
user-specified collection that may be chained or calibration (once
Middlewhere supports queries against calibration collections).
Identifying the collections that hold validity ranges is now handled
automatically. The resulting UI is less error prone, and doesn't
require that the user modify the script as new calib reprocessings
become available.