Remove ap.verify.Dataset use in make_preloaded_export.py.
The use of Dataset -- dating to when I expected one set of central
scripts that would never need to be specialized -- presents a confusing
CLI for users, and required that both the dataset and ap_verify were
set up. The script now takes no arguments and has no preconditions
besides existence of the preloaded repository.